requirements.txt
Flask==2.2.2
Jinja2==3.1.2
python-dotenv==0.21.0